+++ engine/src/flutter/impeller/entity/shaders/gradients/conical_gradient_ssbo_fill.frag
@@ -37,8 +37,8 @@ highp in vec2 v_position;
 out vec4 frag_color;
 
 void main() {
-  int kind = IPComputeConicalKind(frag_info.focus, frag_info.focus_radius,
-                                  frag_info.center, frag_info.radius);
+  float kind = IPComputeConicalKind(frag_info.focus, frag_info.focus_radius,
+                                    frag_info.center, frag_info.radius);
   vec2 res = IPComputeConicalT(kind, frag_info.focus, frag_info.focus_radius,
                                frag_info.center, frag_info.radius, v_position);
